Change names of sceans
