+++ index.html
@@ -6,7 +6,7 @@
   <title>Grab Fruits</title>
 </head>
 <body>
-  <h1>Grab Fruits!</h1>
+  <!-- <h1>Grab Fruits!</h1> -->
   <script src="./dist/index.bundle.js"></script>
 </body>
 </html>
