+++ src/style/style.css
@@ -27,6 +27,7 @@ canvas {
 
 #my-field {
   position: relative;
+  top: 80px;
   display: flex;
   flex-flow: column;
   justify-content: center;
